# openapi.yaml
openapi: 3.0.3
info:
  title: TTSAW (SmartOffice) API
  description: |
    API do zarządzania rezerwacjami sal konferencyjnych na uczelni wyższej.

    ## Funkcjonalności:
    - Autentykacja użytkowników (NextAuth)
    - Przeglądanie dostępnych sal
    - Tworzenie i zarządzanie rezerwacjami
    - Panel administracyjny
    - Kalendarz rezerwacji

    ## Bezpieczeństwo:
    - Rate limiting (5 req/15min dla auth, 20 req/min dla mutations)
    - CSRF protection (Origin/Referer verification)
    - Email validation
    - Max length validation
    - SQL injection protection (prepared statements)

  version: 1.0.0
  contact:
    name: API Support
    email: admin@wydzial.pl
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://ttsaw.production.com
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Session token cookie set by NextAuth
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        retryAfter:
          type: integer
          description: Seconds until retry (for rate limiting)
      required:
        - error
    Room:
      type: object
      properties:
        id:
          type: string
          example: c1a2b3c4d5e6f7
        name:
          type: string
          maxLength: 100
          example: "101"
        building:
          type: string
          example: "A"
        floor:
          type: integer
          example: 1
        capacity:
          type: integer
          example: 30
        equipment:
          type: array
          items:
            type: string
          example: ["projektor", "tablica"]
        description:
          type: string
          maxLength: 500
          nullable: true
          example: "Sala wykładowa z projektorem"
        positionX:
          type: integer
          nullable: true
        positionY:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - building
        - floor
        - capacity
        - equipment
...
paths:
  /api/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: |
        Create a new user account.

        **Rate Limit:** 5 requests per 15 minutes per IP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                  example: student@wydzial.pl
                name:
                  type: string
                  maxLength: 100
                  example: Jan Kowalski
                password:
                  type: string
                  minLength: 6
                  example: securePassword123
              required:
                - email
                - name
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  userId:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      description: Retrieve a list of all rooms with optional filtering
      parameters:
        - name: building
          in: query
          schema:
            type: string
          example: A
        - name: floor
          in: query
          schema:
            type: integer
          example: 1
        - name: minCapacity
          in: query
          schema:
            type: integer
          example: 20
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'

    post:
      tags:
        - Rooms
      summary: Create new room (Admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                building:
                  type: string
                floor:
                  type: integer
                capacity:
                  type: integer
                equipment:
                  type: array
                  items:
                    type: string
                description:
                  type: string
                  maxLength: 500
                positionX:
                  type: integer
                positionY:
                  type: integer
              required:
                - name
                - building
                - floor
                - capacity
      responses:
        '201':
          description: Room created
        '401':
          description: Unauthorized
        '400':
          description: Validation error
...